Introduce start of print replacement
By running the new gcode through austerus-verge again in deposition mode
and adding the head boundaries we have the minimum free area required on
the print bed to do a safe print.

This assumes the gcode gets modified to remain constrained in this space
and all movements outside the designated safe area are above the
calculated safe Z working height.