Use austerus-verge in zmin mode to check new print
One we have found the maximum Z height for all depositions on the plate
we are concerned with where the new gcode tells the print head to go below
this threshold.

This means gcode only passes the test if it moves up above zmin, moves to
the safe print area, then descends there.