Fix large object bug by disabling zsafe check at times
When we are using a planner mode that does not shift or we have already applied
a shift then we should not bail due to printing in the zsafe region as there is
nothing we can do about it and we want to proceed with the print.

After the print the bed will be considered full.